requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
gunicorn==21.2.0